refactor: change node and server for new packet struct

+++ mesh/sources/server.c
@@ -10,7 +10,7 @@ pid_t node_pids[MAX_NODES];
 int client_socket;
 struct sockaddr_in server_address;
 
-void start_node(int node_id)
+void start_node(const int node_id)
 {
     pid_t pid = fork();
     if (pid == 0)
@@ -31,13 +31,21 @@ void start_node(int node_id)
     }
 }
 
-void stop_node(int node_id)
+void stop_node(const int node_id)
 {
-    kill(node_pids[node_id], SIGTERM);
-    waitpid(node_pids[node_id], NULL, 0);
+    if (node_pids[node_id] > 0)
+    {
+        kill(node_pids[node_id], SIGTERM);
+        waitpid(node_pids[node_id], NULL, 0);
+        node_pids[node_id] = 0;
+    }
+    else
+    {
+        log_message("SERVER", MSG_TYPE_ERROR, "Node %d is not running", node_id);
+    }
 }
 
-void handle_signal(int sig)
+void handle_signal(const int sig)
 {
     for (int i = 0; i < MAX_NODES; ++i)
     {
@@ -50,35 +58,33 @@ void handle_signal(int sig)
     exit(EXIT_SUCCESS);
 }
 
-void send_command_to_node(Packet *packet)
+void send_command_to_node(packet_t *packet)
 {
-    if (packet->type == PACKET_TYPE_MAC)
+    if (packet->mac_packet.ttl == 0)
     {
-        if (packet->mac_packet.ttl == 0)
-        {
-            log_message("SERVER", MSG_TYPE_NOT_VALID_DATA, "TTL expired, packet dropped");
-            return;
-        }
+        log_message("SERVER", MSG_TYPE_NOT_VALID_DATA, "TTL expired, packet dropped");
+        return;
+    }
 
-        packet->mac_packet.ttl--;
+    packet->mac_packet.ttl--;
 
-        struct sockaddr_in node_address;
-        node_address.sin_family = AF_INET;
-        node_address.sin_port = htons(CLIENT_BASE_PORT + packet->mac_packet.mac_sender);
-        node_address.sin_addr.s_addr = INADDR_ANY;
+    struct sockaddr_in node_address;
+    node_address.sin_family = AF_INET;
+    node_address.sin_port = htons(CLIENT_BASE_PORT + packet->mac_packet.mac_sender);
+    node_address.sin_addr.s_addr = INADDR_ANY;
 
-        int sent_bytes = sendto(client_socket, packet, sizeof(Packet), 0, (struct sockaddr *)&node_address, sizeof(node_address));
+    int sent_bytes = sendto(client_socket, packet, sizeof(packet_t), 0, (struct sockaddr *)&node_address, sizeof(node_address));
 
-        if (sent_bytes == -1)
-        {
-            log_message("SERVER", MSG_TYPE_ERROR, "sendto() failed");
-        }
+    if (sent_bytes == -1)
+    {
+        log_message("SERVER", MSG_TYPE_ERROR, "sendto() failed");
     }
 }
 
-void create_and_send_command(const int src, const int dest, int graph[MAX_NODES][MAX_NODES], const int size_graph, const char *message)
+void create_and_send_command(const int src, const int dest, int graph[MAX_NODES][MAX_NODES],
+                             const int size_graph, const char *message, const int message_id)
 {
-    Packet packet = create_mac_packet(src, dest, TTL_LIMIT, message);
+    packet_t packet = create_packet(src, dest, TTL_LIMIT, src, dest, message_id, message);
 
     memcpy(packet.network_graph, graph, size_graph * size_graph * sizeof(int));
 
@@ -92,11 +98,13 @@ int main()
     int graph[MAX_NODES][MAX_NODES];
     int matrix_size = 10;
 
-    initialize_graph(MAX_NODES, graph);
-    add_edges(matrix_size, graph);
-
     int distances[MAX_NODES];
     int predecessors[MAX_NODES];
+
+    int message_id = 0;
+
+    initialize_graph(MAX_NODES, graph);
+    add_edges(matrix_size, graph);
 
     client_socket = socket(AF_INET, SOCK_DGRAM, 0);
     if (client_socket == -1)
@@ -117,14 +125,20 @@ int main()
     while (1)
     {
         char command[256];
-        printf("Enter command (send <source_node> <dest_node> <message>): ");
+        printf("Enter command (send <source_node> <dest_node> <message> | stop <node_id>): ");
         fgets(command, sizeof(command), stdin);
 
-        int src_node, dest_node;
+        int src_node, dest_node, node_id;
         char message[MAX_MESSAGE_LENGTH];
         if (sscanf(command, "send %d %d %[^\n]", &src_node, &dest_node, message) == 3)
         {
-            create_and_send_command(src_node, dest_node, graph, MAX_NODES, message);
+            create_and_send_command(src_node, dest_node, graph, MAX_NODES, message, message_id);
+            message_id++;
+        }
+        else if (sscanf(command, "stop %d", &node_id) == 1)
+        {
+            stop_node(node_id);
+            remove_node(node_id, graph);
         }
         else
         {
